refactor(notifications, trending): Organize mock data and update imports
- Moved mock notifications and trending topics data to dedicated files in the src/data/mock directory for better structure and maintainability.
- Updated import paths in the notifications and trending API routes to reflect the new locations of the mock data.
- Removed the old mock data file from the utils directory to streamline the codebase and eliminate redundancy.
